Change PPU to use per-pixel rendering.

+++ src/nes/mapper.rs
@@ -3,12 +3,12 @@ use nes::rom::Rom;
 use nes::mappers::nrom::Nrom;
 
 pub trait Mapper {
-    fn in_range(&self, address: usize) -> bool;
+    fn in_range(&self, address: u16) -> bool;
     fn mirroring(&self) -> MirrorMode;
-    fn read_chr(&self, address: usize) -> u8;
-    fn read_prg(&self, address: usize) -> u8;
-    fn write_chr(&self, address: usize, value: usize);
-    fn write_prg(&self, address: usize, value: usize);
+    fn read_chr(&self, address: u16) -> u8;
+    fn read_prg(&self, address: u16) -> u8;
+    fn write_chr(&self, address: u16, value: u8);
+    fn write_prg(&self, address: u16, value: u8);
 }
 
 pub fn create_mapper(rom: Rom) -> Box<Mapper + Send> {
